requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
azure-core==1.31.0
azure-storage-blob==12.23.1
azure-storage-file-share==12.18.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
easyocr==1.7.2
faiss-cpu==1.8.0.post1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.9
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
PyPDF2==3.0.1
python-bidi==0.6.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
shapely==2.0.6
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
Werkzeug==3.0.4
